move detail header to detail subdirectory
