+++ cpp/include/cudf/fixed_point/detail/floating_conversion.hpp
@@ -26,14 +26,6 @@
 #include <cstring>
 
 namespace CUDF_EXPORT numeric {
-
-/**
- * @addtogroup floating_conversion
- * @{
- * @file
- * @brief fixed_point <--> floating-point conversion functions.
- */
-
 namespace detail {
 
 /**
@@ -1141,6 +1133,4 @@ CUDF_HOST_DEVICE inline FloatingType convert_integral_to_floating(Rep const& val
 }
 
 }  // namespace detail
-
-/** @} */  // end of group
 }  // namespace CUDF_EXPORT numeric
